refactor(rag): restructure TUI into src/rag/tui package
- Move all TUI logic (app, widgets, constants) from scripts/retrieve.py
  into src/rag/tui/ following the existing Strategy/module pattern
- scripts/retrieve.py is now a thin entry point (~10 lines)
- Move textual from runtime dependencies to optional extra [tui] and
  dev dependency group; not needed by the ingestion pipeline itself
- Add textual to pre-commit mypy additional_dependencies so both
  environments (Docker make typecheck and pre-commit hook) use stubs
- Replace tests/test_retrieve.py (script-level, path-hacking) with
  proper tests/tui/ covering src/rag/tui modules
