+++ .secrets.baseline
@@ -152,15 +152,6 @@
         "line_number": 11
       }
     ],
-    "tests/test_backup.py": [
-      {
-        "type": "Secret Keyword",
-        "filename": "tests/test_backup.py",
-        "hashed_secret": "a62f2225bf70bfaccbc7f1ef2a397836717377de",
-        "is_verified": false,
-        "line_number": 174
-      }
-    ],
     "tests/test_rag.py": [
       {
         "type": "Secret Keyword",
@@ -172,5 +163,5 @@
       }
     ]
   },
-  "generated_at": "2026-04-25T22:25:56Z"
+  "generated_at": "2026-04-25T22:40:00Z"
 }
