+++ src/rag/tui/__init__.py
@@ -1,0 +1,5 @@
+"""Textual TUI package for interactive RAG retrieval evaluation."""
+
+from rag.tui.app import RetrievalApp
+
+__all__ = ["RetrievalApp"]
